refactor(#47): stats 도메인 상태 관리 세팅
